perf: variable select component
