+++ projects/app/src/components/core/module/Flow/components/nodes/abandon/NodeVariable.tsx
@@ -11,7 +11,7 @@ import type { VariableItemType } from '@fastgpt/global/core/module/type.d';
 import MyIcon from '@/components/Icon';
 import { customAlphabet } from 'nanoid';
 const nanoid = customAlphabet('abcdefghijklmnopqrstuvwxyz1234567890', 6);
-import VariableEditModal, { addVariable } from '../../../../VariableEditModal';
+import VariableEditModal, { addVariable } from '../../modules/VariableEdit';
 import { onChangeNode } from '../../../FlowProvider';
 
 export const defaultVariable: VariableItemType = {
